import ty's constants from 2024-improvements branch
